Move glue.qt.widget_properties to glue.utils.qt.widget_properties
