+++ glue/qt/widgets/subset_facet.py
@@ -6,7 +6,7 @@ from matplotlib import cm
 from glue.external.qt import QtGui
 from glue.core.util import colorize_subsets, facet_subsets
 from glue.qt.qtutil import load_ui
-from glue.qt.widget_properties import (ButtonProperty, FloatLineProperty,
+from glue.utils.qt.widget_properties import (ButtonProperty, FloatLineProperty,
                                        ValueProperty)
 from glue.utils.array import pretty_number
 from glue.utils import Pointer
