+++ glue/utils/qt/tests/test_widget_properties.py
